+++ batch/src/main/java/store/_0982/batch/domain/product/ProductRepository.java
@@ -7,6 +7,4 @@ import java.util.Optional;
 import java.util.UUID;
 
 public interface ProductRepository {
-    Optional<Product> findById(UUID productId);
-    List<Product> findAllByIdIn(List<UUID> ids);
 }
