+++ batch/src/main/java/store/_0982/batch/infrastructure/grouppurchase/GroupPurchaseRepositoryAdaptor.java
@@ -26,63 +26,7 @@ public class GroupPurchaseRepositoryAdaptor implements GroupPurchaseRepository {
     }
 
     @Override
-    public Optional<GroupPurchase> findById(UUID groupPurchaseId) {
-        return groupPurchaseJpaRepository.findById(groupPurchaseId);
-    }
-
-    @Override
-    public Page<GroupPurchase> findAll(Pageable pageable) {
-        return groupPurchaseJpaRepository.findAll(pageable);
-    }
-
-    @Override
-    public Page<GroupPurchase> findAllBySellerId(UUID sellerId, Pageable pageable) {
-        return groupPurchaseJpaRepository.findAllBySellerId(sellerId, pageable);
-    }
-
-    @Override
-    public void delete(GroupPurchase groupPurchase) {
-        groupPurchaseJpaRepository.delete(groupPurchase);
-    }
-
-    @Override
-    public GroupPurchase saveAndFlush(GroupPurchase groupPurchase) {
-        return groupPurchaseJpaRepository.saveAndFlush(groupPurchase);
-    }
-
-    @Override
     public List<GroupPurchase> saveAll(List<GroupPurchase> groupPurchaseList) {
         return groupPurchaseJpaRepository.saveAll(groupPurchaseList);
-    }
-
-    @Override
-    @Transactional
-    public int openReadyGroupPurchases(OffsetDateTime now) {
-        return groupPurchaseJpaRepository.openReadyGroupPurchases(now);
-    }
-
-    @Override
-    public boolean existsByProductId(UUID productId) {
-        return groupPurchaseJpaRepository.existsByProductId((productId));
-    }
-
-    @Override
-    public boolean existsByProductIdAndStatusIn(UUID productId, List<GroupPurchaseStatus> statuses) {
-        return groupPurchaseJpaRepository.existsByProductIdAndStatusIn(productId, statuses);
-    }
-
-    @Override
-    public List<GroupPurchase> findAllByStatusAndStartDateBefore(GroupPurchaseStatus status, OffsetDateTime now) {
-        return groupPurchaseJpaRepository.findAllByStatusAndStartDateBefore(status, now);
-    }
-
-    @Override
-    public List<GroupPurchase> findAllByGroupPurchaseIdIn(List<UUID> groupPurchaseIds) {
-        return groupPurchaseJpaRepository.findAllByGroupPurchaseIdIn(groupPurchaseIds);
-    }
-
-    @Override
-    public void markAsSettled(List<UUID> uuids) {
-        groupPurchaseJpaRepository.markAsSettled(uuids, OffsetDateTime.now());
     }
 }
