+++ batch/src/main/java/store/_0982/batch/infrastructure/product/ProductRepositoryAdaptor.java
@@ -15,14 +15,4 @@ public class ProductRepositoryAdaptor implements ProductRepository {
 
     private final ProductJpaRepository productJpaRepository;
 
-    public Optional<Product> findById(UUID productId) {
-        return productJpaRepository.findById(productId);
-    }
-
-    @Override
-    public List<Product> findAllByIdIn(List<UUID> ids) {
-        return productJpaRepository.findAllById(ids);
-    }
-
-
 }
