+++ commerce/src/main/java/store/_0982/commerce/application/grouppurchase/GroupPurchaseService.java
@@ -10,12 +10,11 @@ import store._0982.commerce.application.grouppurchase.dto.*;
 import store._0982.commerce.application.grouppurchase.event.GroupPurchaseCreatedEvent;
 import store._0982.commerce.application.grouppurchase.event.GroupPurchaseDeletedEvent;
 import store._0982.commerce.application.grouppurchase.event.GroupPurchaseUpdatedEvent;
-import store._0982.commerce.domain.grouppurchase.GroupPurchase;
 import store._0982.commerce.domain.grouppurchase.GroupPurchaseRepository;
-import store._0982.commerce.domain.grouppurchase.GroupPurchaseStatus;
 import store._0982.commerce.domain.product.ProductRepository;
 import store._0982.commerce.exception.CustomErrorCode;
 import store._0982.commerce.infrastructure.client.member.MemberClient;
+import store._0982.common.domain.grouppurchase.*;
 import store._0982.common.domain.product.Product;
 import store._0982.common.dto.PageResponse;
 import store._0982.common.exception.CustomException;
