+++ commerce/src/main/java/store/_0982/commerce/application/grouppurchase/dto/GroupPurchaseDetailInfo.java
@@ -1,7 +1,7 @@
 package store._0982.commerce.application.grouppurchase.dto;
 
-import store._0982.commerce.domain.grouppurchase.GroupPurchase;
-import store._0982.commerce.domain.grouppurchase.GroupPurchaseStatus;
+import store._0982.common.domain.grouppurchase.GroupPurchase;
+import store._0982.common.domain.grouppurchase.GroupPurchaseStatus;
 import store._0982.common.domain.product.ProductCategory;
 
 import java.time.OffsetDateTime;
