+++ commerce/src/main/java/store/_0982/commerce/application/order/OrderCommandService.java
@@ -21,13 +21,13 @@ import store._0982.commerce.application.order.event.OrderCartCompletedEvent;
 import store._0982.commerce.application.product.ProductService;
 import store._0982.commerce.application.settlement.OrderSettlementService;
 import store._0982.commerce.domain.cart.Cart;
-import store._0982.commerce.domain.grouppurchase.GroupPurchase;
 import store._0982.commerce.domain.order.Order;
 import store._0982.commerce.domain.order.OrderCancellationPolicy;
 import store._0982.commerce.domain.order.OrderRepository;
 import store._0982.commerce.exception.CustomErrorCode;
 import store._0982.commerce.infrastructure.client.member.MemberClient;
 import store._0982.commerce.infrastructure.client.member.dto.ProfileInfo;
+import store._0982.common.domain.grouppurchase.GroupPurchase;
 import store._0982.common.domain.order.OrderStatus;
 import store._0982.common.dto.ResponseDto;
 import store._0982.common.exception.CustomException;
