+++ commerce/src/main/java/store/_0982/commerce/application/order/OrderQueryService.java
@@ -12,7 +12,6 @@ import store._0982.commerce.application.order.dto.OrderCancelInfo;
 import store._0982.commerce.application.order.dto.OrderDetailInfo;
 import store._0982.commerce.application.order.dto.OrderInfo;
 import store._0982.commerce.application.product.dto.OrderVectorInfo;
-import store._0982.commerce.domain.grouppurchase.GroupPurchase;
 import store._0982.commerce.domain.grouppurchase.GroupPurchaseRepository;
 import store._0982.commerce.domain.order.Order;
 import store._0982.commerce.domain.order.OrderRepository;
@@ -20,6 +19,7 @@ import store._0982.commerce.domain.product.ProductRepository;
 import store._0982.commerce.domain.product.ProductVector;
 import store._0982.commerce.exception.CustomErrorCode;
 import store._0982.commerce.infrastructure.product.ProductVectorJpaRepository;
+import store._0982.common.domain.grouppurchase.GroupPurchase;
 import store._0982.common.domain.order.OrderStatus;
 import store._0982.common.domain.product.Product;
 import store._0982.common.dto.PageResponse;
