+++ commerce/src/main/java/store/_0982/commerce/application/product/ProductService.java
@@ -10,10 +10,10 @@ import org.springframework.transaction.annotation.Transactional;
 import store._0982.commerce.application.product.dto.*;
 import store._0982.commerce.application.product.event.ProductUpsertedEvent;
 import store._0982.commerce.domain.grouppurchase.GroupPurchaseRepository;
-import store._0982.commerce.domain.grouppurchase.GroupPurchaseStatus;
 import store._0982.commerce.domain.product.ProductRepository;
 import store._0982.commerce.domain.product.ProductVectorRepository;
 import store._0982.commerce.exception.CustomErrorCode;
+import store._0982.common.domain.grouppurchase.GroupPurchaseStatus;
 import store._0982.common.domain.product.Product;
 import store._0982.common.dto.PageResponse;
 import store._0982.common.exception.CustomException;
