--- commerce/src/main/java/store/_0982/commerce/domain/grouppurchase/GroupPurchase.java
@@ -1,205 +1,0 @@
-package store._0982.commerce.domain.grouppurchase;
-
-import jakarta.persistence.*;
-import lombok.AccessLevel;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-import org.hibernate.annotations.CreationTimestamp;
-import org.hibernate.annotations.UpdateTimestamp;
-import store._0982.commerce.exception.CustomErrorCode;
-import store._0982.common.exception.CustomException;
-import store._0982.common.kafka.dto.GroupPurchaseEvent;
-
-import java.time.OffsetDateTime;
-import java.util.UUID;
-
-@Getter
-@NoArgsConstructor(access = AccessLevel.PROTECTED)
-@Entity
-@Table(name = "\"group_purchase\"", schema = "product_schema")
-public class GroupPurchase {
-    @Id
-    private UUID groupPurchaseId;
-
-    @Version
-    @Column(name = "version")
-    private Long version;
-
-    @Column(name = "min_quantity", nullable = false)
-    private int minQuantity;
-
-    @Column(name = "max_quantity")
-    private int maxQuantity;
-
-    @Column(name = "title", nullable = false, length = 100)
-    private String title;
-
-    @Column(name = "description")
-    private String description;
-
-    @Column(name = "discounted_price", nullable = false)
-    private Long discountedPrice;
-
-    @Enumerated(EnumType.STRING)
-    @Column(name = "status", nullable = false)
-    private GroupPurchaseStatus status;
-
-    @Column(name = "start_date", nullable = false)
-    private OffsetDateTime startDate;
-
-    @Column(name = "end_date", nullable = false)
-    private OffsetDateTime endDate;
-
-    @Column(name = "seller_id", nullable = false)
-    private UUID sellerId;
-
-    @Column(name = "product_id", nullable = false)
-    private UUID productId;
-
-    @Column(name = "image_url", length = 512)
-    private String imageUrl;
-
-    @Column(name = "current_quantity", nullable = false)
-    private int currentQuantity = 0;
-
-    @Column(name = "returned_at")
-    private OffsetDateTime returnedAt;
-
-    @Column(name = "succeeded_at")
-    private OffsetDateTime succeededAt;
-
-    @Column(name = "settled_at")
-    private OffsetDateTime settledAt; // 정산 완료 시간
-
-    @Column(name = "created_at", nullable = false)
-    @CreationTimestamp
-    private OffsetDateTime createdAt;
-
-    @Column(name = "updated_at")
-    @UpdateTimestamp
-    private OffsetDateTime updatedAt;
-    
-    public GroupPurchase(int mintQuantity,
-                         int maxQuantity,
-                         String title,
-                         String description,
-                         Long discountedPrice,
-                         OffsetDateTime startDate,
-                         OffsetDateTime endDate,
-                         UUID sellerId,
-                         UUID productId,
-                         String imageUrl){
-        this.groupPurchaseId = UUID.randomUUID();
-        this.minQuantity = mintQuantity;
-        this.maxQuantity = maxQuantity;
-        this.title = title;
-        this.description = description;
-        this.discountedPrice = discountedPrice;
-        this.status = GroupPurchaseStatus.SCHEDULED;
-        this.startDate = startDate;
-        this.endDate = endDate;
-        this.sellerId = sellerId;
-        this.productId = productId;
-        this.imageUrl = imageUrl;
-        this.currentQuantity = 0;
-    }
-
-    private boolean canParticipate(int quantity) {
-        return status == GroupPurchaseStatus.OPEN
-                && (this.currentQuantity + quantity <= this.maxQuantity);
-    }
-
-    private void checkAndUpdateStatusIfMaxReached() {
-        if (this.currentQuantity == this.maxQuantity) {
-            this.status = GroupPurchaseStatus.SUCCESS;
-        }
-    }
-
-    public boolean increaseQuantity(int quantity) {
-        if (!canParticipate(quantity)) {
-            return false;
-        }
-
-        this.currentQuantity += quantity;
-        checkAndUpdateStatusIfMaxReached();
-
-        return true;
-    }
-
-    public void decreaseQuantity(int quantity){
-        if(this.currentQuantity - quantity < 0){
-            throw new CustomException(CustomErrorCode.DECREASE_QUANTITY_FAILED);
-        }
-        this.currentQuantity -= quantity;
-    }
-
-    public boolean isInReversedPeriod() {
-        if (this.succeededAt == null) {
-            return false;
-        }
-        return OffsetDateTime.now().isBefore(this.succeededAt.plusDays(2));
-    }
-
-    public boolean isInReturnedPeriod() {
-        if (this.succeededAt == null)
-            return false;
-        return OffsetDateTime.now().isAfter(this.succeededAt.plusDays(2))
-                && OffsetDateTime.now().isBefore(this.succeededAt.plusWeeks(2));
-    }
-
-    public void updateGroupPurchase(int mintQuantity,
-                                    int maxQuantity,
-                                    String title,
-                                    String description,
-                                    Long discountedPrice,
-                                    OffsetDateTime startDate,
-                                    OffsetDateTime endDate,
-                                    UUID productId,
-                                    String imageUrl){
-        this.minQuantity = mintQuantity;
-        this.maxQuantity = maxQuantity;
-        this.title = title;
-        this.description = description;
-        this.discountedPrice = discountedPrice;
-        this.startDate = startDate;
-        this.endDate = endDate;
-        this.productId = productId;
-        this.imageUrl = imageUrl;
-    }
-  
-    public void updateStatus(GroupPurchaseStatus status){
-        this.status = status;
-    }
-
-    public void markAsReturned() {
-        if (this.returnedAt != null) {
-            throw new IllegalStateException("이미 환불 처리된 공동구매입니다.");
-        }
-        this.returnedAt = OffsetDateTime.now();
-    }
-
-    public boolean isReturned() {
-        return this.returnedAt != null;
-    }
-
-    public GroupPurchaseEvent toEvent(GroupPurchaseEvent.Status groupPurchaseStatus,
-                                      GroupPurchaseEvent.EventStatus kafkaStatus,
-                                      Long originalPrice,
-                                      GroupPurchaseEvent.ProductCategory category) {
-        return new GroupPurchaseEvent(
-                this.groupPurchaseId,
-                this.sellerId,
-                this.title,
-                this.description,
-                this.discountedPrice,
-                this.productId,
-                groupPurchaseStatus,
-                this.endDate.toString(),
-                this.updatedAt.toString(),
-                this.currentQuantity,
-                kafkaStatus,
-                originalPrice,
-                category
-        );
-    }
-}
