+++ common/src/main/java/store/_0982/common/exception/EntityErrorCode.java
@@ -9,7 +9,9 @@ import org.springframework.http.HttpStatus;
 public enum EntityErrorCode implements ErrorCode {
 
     INVALID_SETTLEMENT_AMOUNT(HttpStatus.BAD_REQUEST, "잘못된 정산 금액입니다."),
-    INSUFFICIENT_BALANCE(HttpStatus.BAD_REQUEST, "정산 잔액이 부족합니다.");
+    INSUFFICIENT_BALANCE(HttpStatus.BAD_REQUEST, "정산 잔액이 부족합니다."),
+
+    DECREASE_QUANTITY_FAILED(HttpStatus.BAD_REQUEST, "공동구매 수량 감소에 실패했습니다.");
 
     private final HttpStatus httpStatus;
     private final String message;
